feat: Actuator screens implemented
These screens allows the user to manage the actuators of the system.

+++ app/src/main/java/com/arejas/dashboardofthings/domain/entities/database/Actuator.java
@@ -57,9 +57,9 @@ public class Actuator implements Parcelable {
 
     private String dataUnit;
 
-    private Long locationLat;
+    private Double locationLat;
 
-    private Long localtionLong;
+    private Double localtionLong;
 
     private Boolean showInMainDashboard;
 
@@ -105,12 +105,12 @@ public class Actuator implements Parcelable {
         if (in.readByte() == 0) {
             locationLat = null;
         } else {
-            locationLat = in.readLong();
+            locationLat = in.readDouble();
         }
         if (in.readByte() == 0) {
             localtionLong = null;
         } else {
-            localtionLong = in.readLong();
+            localtionLong = in.readDouble();
         }
         showInMainDashboard = in.readByte() != 0;
     }
@@ -255,19 +255,19 @@ public class Actuator implements Parcelable {
         this.dataUnit = sDataUnit;
     }
 
-    public Long getLocationLat() {
+    public Double getLocationLat() {
         return locationLat;
     }
 
-    public void setLocationLat(Long lLocationLat) {
+    public void setLocationLat(Double lLocationLat) {
         this.locationLat = lLocationLat;
     }
 
-    public Long getLocaltionLong() {
+    public Double getLocaltionLong() {
         return localtionLong;
     }
 
-    public void setLocaltionLong(Long lLocaltionLong) {
+    public void setLocaltionLong(Double lLocaltionLong) {
         this.localtionLong = lLocaltionLong;
     }
 
@@ -330,13 +330,13 @@ public class Actuator implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeLong(locationLat);
+            parcel.writeDouble(locationLat);
         }
         if (localtionLong == null) {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeLong(localtionLong);
+            parcel.writeDouble(localtionLong);
         }
         parcel.writeByte((byte) (showInMainDashboard ? 1 : 0));
     }
